users_list route, backend and html implemented (links don't work yet

+++ sample_data/populate-user.sql
@@ -5,7 +5,7 @@ CREATE TABLE users (
     username varchar(20),
     email varchar(50),
     password char(60),
-    reputation bigint
+    reputation integer
 );
 
 INSERT INTO users
